feat(stdlib): implement zero-cost error return traces in runtime and stdlib (ERR-TRACE-1)
- Expose ori.err_trace stdlib module (push, format) in stdlib/err_trace.orl
- Wire ori_err_trace_push (with i64 line) and ori_err_trace_format in native and C backends
- Stage updated native runtime staticlib and cdylib
- Add E2E tests for native execution and C backend generation in multifile_imports.rs
- Document in spec 12, BACKLOG.md, and CHANGELOG.md

+++ compiler/crates/ori-runtime/src/lib.rs
@@ -1924,7 +1924,7 @@ pub unsafe extern "C" fn ori_reactor_wake() {
 #[no_mangle]
 pub unsafe extern "C" fn ori_err_trace_push(
     file: *const std::ffi::c_char,
-    line: u32,
+    line: i64,
     err_str: *mut u8,
 ) -> *mut u8 {
     let file_str = if file.is_null() {
